pyproject.toml
[project]
name = "datastream-mysql-and-clickhouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.1",
    "clickhouse-connect>=0.8.17",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=1.4.54",
]
